Fix bleeding of -Wno-constant-conversion and -Wno-shorten-64-to-32
The -Wno-constant-conversion and -Wno-shorten-64-to-32 flags were
bleeding into //chrome/browser:browser as per
`gn desc out/Default //chrome/browser:browser cflags --blame` because
they were part of the public config of flatbuffers. This CL removes them
from the public config.

Fixed: 513276104
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7852315
Commit-Queue: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1632813}

+++ components/autofill/core/browser/suggestions/payments/credit_card_suggestion_generator_unittest.cc
@@ -866,7 +866,7 @@ TEST_F(CreditCardSuggestionGeneratorTest,
   const base::Time kNow = AutofillClock::Now();
   const base::Time kDisuseTime =
       kNow - kDisusedDataModelTimeDelta - base::Days(1);
-  size_t card_number = 4111111111111111ul;
+  uint64_t card_number = 4111111111111111ULL;
 
   std::vector<CreditCard> credit_cards;
   for (bool is_local : {false, true}) {
